# Conflicts:
#	src/main/java/net/matez/wildnature/common/objects/features/FeatureRegistry.java
#	src/main/java/net/matez/wildnature/common/objects/features/ModWorldEvents.java
#	src/main/java/net/matez/wildnature/common/objects/features/WNCombinedFeature.java
#	src/main/java/net/matez/wildnature/common/objects/features/WNTreeFeature.java
#	src/main/java/net/matez/wildnature/setup/WildNature.java
#	src/main/resources/META-INF/mods.toml

+++ src/main/java/net/matez/wildnature/common/objects/blocks/ores/OreDropConfig.java
@@ -12,7 +12,7 @@ import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.state.BlockState;
 
 public interface OreDropConfig {
-    ItemStack getDrops(BlockState state, ServerLevel level, float luck);
+    ItemStack getDrops(BlockState state, ServerLevel level, int luck);
 
     int getExpDrops(BlockState state, LevelReader reader, int fortune, int silkTouch);
 }
